requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pytest==8.0.0
httpx==0.26.0
python-dotenv==1.0.1 